Feature/backend crawling controller

See merge request s05-final/S05P31A402!35

+++ backend/src/main/java/com/ssacretary/db/repository/SettingRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface SettingRepository extends JpaRepository<Setting,Integer> {
-    Setting findBySettingId(Long settingId);
+    Setting findBySettingId(int settingId);
 
     List<Setting> findByUser_Email(String email);
